+++ src/net/ip4.c
@@ -56,7 +56,7 @@ uint32_t stored_fragments = 0;
 
 static void prepare_packet_to_send(ip4_header_t* packet) {
 	packet->version = 4;
-	packet->id = (uint16_t)(pit_getTickNum() % 65535);
+	packet->id = ((uint32_t)pit_getTickNum() % 65535);
 	packet->checksum = 0;
 	packet->checksum = endian_swap16(net_calculate_checksum((uint8_t*)packet, sizeof(ip4_header_t), 0));
 }
